+++ cookbooks/aws-parallelcluster-environment/attributes/environment.rb
@@ -2,6 +2,10 @@ default['cluster']['stack_name'] = nil
 
 default['cluster']['proxy'] = 'NONE'
 
+# Override NFS service ports that fall within the Linux ephemeral port range (32768-60999)
+default['nfs']['port']['lockd'] = 32_763
+default['nfs']['port']['rquotad'] = 32_764
+
 # For performance, set NFS threads to min(256, max(8, num_cores * 4))
 default['cluster']['nfs']['threads'] = [[node['cpu']['cores'].to_i * 4, 8].max, 256].min